Web-Scraping-Leboncoin-main/requirements.txt
attrs==25.3.0
backoff==2.2.1
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
flasgger==0.9.7.1
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.59.0
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
loguru==0.7.3
MarkupSafe==3.0.2
matplotlib==3.10.3
mistune==3.1.3
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
playwright==1.54.0
psycopg2-binary==2.9.10
pyee==13.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
scrapfly-sdk==0.8.23
six==1.17.0
tenacity==9.1.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
PyJWT==2.10.1
beautifulsoup4==4.13.4 
soupsieve==2.7
